Make sure random_frame_offset is `int`

+++ src/data_loaders.py
@@ -49,8 +49,8 @@ def data_gen_standard(id, audio_repr_path, gt, pack):
         if sampling == 'random':
             for i in range(0, param_sampling):
                 random_uniform = np.random.random()
-                random_frame_offset = round(
-                    random_uniform * (frames_num - config['xInput']))
+                random_frame_offset = int(round(
+                    random_uniform * (frames_num - config['xInput'])))
 
                 # idx * bands * bytes per float
                 offset = random_frame_offset * config['yInput'] * 2
@@ -111,8 +111,8 @@ def data_gen_feature_combination(id, audio_repr_path, gt, pack):
         if sampling == 'random':
             for i in range(0, param_sampling):
                 random_uniform = np.random.random()
-                random_frame_offset = round(
-                    random_uniform * (frames_num - config['xInput']))
+                random_frame_offset = int(round(
+                    random_uniform * (frames_num - config['xInput'])))
 
                 x = np.hstack([read_mmap(path,
                                          config['xInput'],
